requirements.txt
psycopg2-binary
jupyterlab
ipython-sql
pandas
boto3
s3fs